Reduce maintenance feature signal tests
Split the oversized mounted maintenance signal spec into focused update and ESP flash owners, keep only high-value mounted flows, and move cancel-refresh checks to workflow specs.
