+++ tools/dev/oversized_test_allowlist.yml
@@ -4,7 +4,5 @@
 threshold_lines: 700
 report_limit: 10
 allowlist:
-  # Maintenance mounted-DOM coverage now lives in one Node-side regression owner instead of four legacy bridge specs.
-  apps/ui/tests/maintenance_feature_signal.spec.ts: Mounted maintenance regression matrix intentionally stays together while replacing the deleted legacy createPanel Playwright split.
   # Cars browser regression keeps the moved car-selection matrix outside smoke.
   apps/ui/tests/regression.cars.spec.ts: Cars browser regression matrix moved out of smoke but intentionally stays together across library selection, manual entry, and wizard flows.
